requirements.txt
fastapi
uvicorn
python-socketio
google-generativeai
